chore(reorg): Moved moons.ts to its associated Svelte folder
